Modification du processus de connection
