+++ app/src/main/java/com/etna/mob4/entities/Messages.kt
@@ -2,6 +2,8 @@ package com.etna.mob4.entities
 
 /**
  * Created by kumatetsu on 12/07/2018.
+ *
+ * Data class for messages
  */
 
 data class Message(
