+++ app/src/main/res/layout/main_layout.xml
@@ -6,19 +6,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <Button
-        android:id="@+id/create"
-        style="@android:style/Widget.Material.Light.Button.Borderless.Colored"
-        android:layout_width="153dp"
-        android:layout_height="45dp"
-        android:layout_marginBottom="40dp"
-        android:layout_marginEnd="16dp"
-        android:onClick="goToCreateUser"
-        android:text="Create account "
-        android:visibility="visible"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
     <TextView
         android:id="@+id/textView"
         android:layout_width="305dp"
@@ -35,15 +22,15 @@
     <Button
         android:id="@+id/connect"
         style="@android:style/Widget.Material.Light.Button.Borderless.Colored"
-        android:layout_width="159dp"
+        android:layout_width="138dp"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="40dp"
-        android:layout_marginEnd="32dp"
-        android:layout_marginStart="16dp"
-        android:onClick="goToConnectUser"
-        android:text="Connect"
+        android:layout_marginBottom="16dp"
+        android:layout_marginStart="246dp"
+        android:onClick="disconnect"
+        android:text="Disconnect"
+        android:textSize="10sp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/create"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_chainStyle="spread_inside"
         app:layout_constraintStart_toStartOf="parent" />
