+++ app/src/main/res/layout/sign_in_layout.xml
@@ -8,16 +8,17 @@
 
     <Button
         android:id="@+id/button"
-        style="@android:style/Widget.Material.Light.Button.Borderless.Small"
+        style="@android:style/Widget.Material.Light.Button.Borderless.Colored"
         android:layout_width="179dp"
         android:layout_height="44dp"
-        android:layout_marginBottom="33dp"
+        android:layout_marginBottom="59dp"
         android:layout_marginEnd="148dp"
         android:layout_marginStart="148dp"
         android:onClick="connect"
         android:text="Connect !"
-        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintBottom_toTopOf="@+id/button4"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.494"
         app:layout_constraintStart_toStartOf="parent" />
 
     <EditText
@@ -73,5 +74,20 @@
         app:layout_constraintBottom_toTopOf="@+id/user_pswd"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/user_email" />
+
+    <Button
+        android:id="@+id/button4"
+        style="@android:style/Widget.Material.Light.Button.Borderless.Colored"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="16dp"
+        android:layout_marginEnd="144dp"
+        android:layout_marginStart="145dp"
+        android:onClick="goToCreateUser"
+        android:text="create account"
+        android:textSize="8sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 
 </android.support.constraint.ConstraintLayout>
